factor out Linter struct
